Extract common code in CalendarActivity and GroupDetailsActivity into extension methods
